tests/wpt/web-platform-tests/fullscreen/api/element-request-fullscreen-options.tentative.html

<!DOCTYPE html>
<title>Element#requestFullscreen({ screen }) tentative support</title>
<link rel="help" href="https://w3c.github.io/window-placement/" />
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
    promise_test(async (t) => {
        await test_driver.set_permission(
            { name: "window-placement" },
            "granted"
        );

        const options = {
            get screen() {
                return undefined;
            },
        };

        await promise_rejects_js(
            t,
            TypeError,
            document.body.requestFullscreen(options)
        );
    }, "fullscreenOptions.screen getter is invoked on requestFullscreen");
</script>
